Add navigation menu with routing
